Add initial project files
This commit introduces several foundational files for the project. It includes the creation of TypeScript and CSS files for the basic structure of the application, as well as configuration files for ESLint, gitignore, package.json and package-lock.json. Preliminary setup for Redux and Vite can also be observed in these files, as well as elements for product fetching.